Made /create-chat and /add-user-to-chat (#2)
* wip

* Made /create-chat and /add-user-to-chat

+++ server/internals/dbMigrations/tables/ChatToUser.cs
@@ -2,9 +2,9 @@
 {
     public class ChatToUser
     {
-        public int UserID { get; set; }
+        public ulong UserID { get; set; }
         public User User { get; set; }
-        public int ChatID { get; set; }
+        public ulong ChatID { get; set; }
         public Chat Chat { get; set; }
     }
 }
