+++ server/internals/dbMigrations/tables/ProfilePicture.cs
@@ -5,7 +5,7 @@ namespace server.internals.dbMigrations.tables
     public class ProfilePicture
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int UserID { get; set; }
+        public ulong UserID { get; set; }
         public string Base64EncodedImage { get; set; }
         public User User { get; set; }
     }
