+++ server/internals/dbMigrations/tables/WatchedBy.cs
@@ -2,9 +2,9 @@
 {
     public class WatchedBy
     {
-        public int MessageID { get; set; }
+        public ulong MessageID { get; set; }
         public Message Message { get; set; }
-        public int UserID { get; set; }
+        public ulong UserID { get; set; }
         public User User { get; set; }
     }
 }
